backend/requirements.txt
aniso8601==9.0.1
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
blinker==1.5
boto3==1.26.61
botocore==1.29.61
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
email-validator==1.3.1
executing==1.2.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.1
idna==3.4
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.0
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
WTForms==3.0.1
